feat(routes): execute the sweep, and chain check:audience into the drift gate
The map (#177) decided; this executes. check:audience is green and is now
in both check:drift and check:drift:ci, so the next route that ships
without an audience fails the build.

One telling (#182). /story keeps its URL and takes /story/road's content.
/story/road, /about and /the-work redirect to it. The three led with the
same sentence rearranged: 'Made by community', 'Built with communities,
not for them', 'Led by community'.

One deck (#183). /pitch and /pitch/deck redirect to /pitch/road;
/pitch/control-room is repointed off /pitch/deck, which stops existing.
Both rendered the same 16 slides from the same module.

Four internal tools move behind the admin gate: investor-lab,
pitch-workshop, miro-board, deck-photo-review. Two of them said so in
their own source - 'Internal working surface, not a funder destination.
Indexed by accident until 2026-07-26.' noindex was never a gate. The
photo board takes a distinct name because /admin/photo-review already
exists as an older redirect stub.

Retired, each having earned it (#184): /shop/[slug], which queried
Supabase while products.ts is canon and resolved nothing; /media/page.tsx,
shadowed by its own config redirect and unreachable, open since ruling S;
the four /design/* mocks, already 404ing in production.

/shop/stretch-bed now resolves instead of 404ing. The canonical slug
redirects to the live URL rather than the live URL moving, because
stretch-bed-single sits in Stripe checkout flows.

/admin/products executed at last: marked retire 2026-07-20 in
route-review.ts and never swept, thirteen days a live page reading a data
source that contradicts canon. Six orphaned legacy-page.tsx files deleted
with it - those seven files are why the route count read 209 not 202.

Two guards caught real things while this landed. The admin route directory
(PR #168) failed on /admin/products as a ghost entry. And assertion 16
flagged ten redirects that the block comment above them already
documented, which was the assertion being too literal, not the config
being wrong: it now reads comment runs, so a section comment covering a
group counts.

deck.guards.test.ts drops deck-public.tsx from PUBLIC_RENDERERS in the
same commit that deletes it, as ticket #183 required.

tsc clean, 420 tests, build clean, check:drift:ci, check:voice,
check:content-gate, check:audience all pass.
