+++ v2/package.json
@@ -19,8 +19,8 @@
     "check:artifacts": "node scripts/check-artifact-drift.mjs",
     "check:qbe-readiness": "node scripts/check-qbe-readiness.mjs",
     "check:story-coverage": "node scripts/check-story-coverage.mjs",
-    "check:drift": "npm run check:community-copy && npm run check:qbe-guardrails && npm run check:canon && npm run check:artifacts && npm run check:qbe-readiness && npm run check:story-coverage && npm run check:asset-drift && npm run check:register && npm run check:retired-figures && npm run check:voice",
-    "check:drift:ci": "npm run check:community-copy && npm run check:qbe-guardrails && npm run check:canon && npm run check:artifacts && npm run check:qbe-readiness && npm run check:story-coverage && npm run check:retired-figures && npm run check:voice",
+    "check:drift": "npm run check:community-copy && npm run check:qbe-guardrails && npm run check:canon && npm run check:artifacts && npm run check:qbe-readiness && npm run check:story-coverage && npm run check:asset-drift && npm run check:register && npm run check:audience && npm run check:retired-figures && npm run check:voice",
+    "check:drift:ci": "npm run check:community-copy && npm run check:qbe-guardrails && npm run check:canon && npm run check:artifacts && npm run check:qbe-readiness && npm run check:story-coverage && npm run check:audience && npm run check:retired-figures && npm run check:voice",
     "funding:sweep": "node scripts/funding-sweep.mjs",
     "audit:model": "node --env-file=.env.local scripts/audit-data-model.mjs",
     "sync:el": "python3 scripts/sync-goods-el.py",
