+++ v2/src/app/admin/investor-lab/investor-lab-client.tsx
