+++ v2/src/app/admin/investor-lab/page.tsx
