--- v2/src/app/design/layout.tsx
@@ -1,9 +1,0 @@
-import type { Metadata } from 'next';
-
-export const metadata: Metadata = {
-  robots: { index: false },
-};
-
-export default function DesignLayout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
-}
