--- v2/src/app/mission/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default function MissionPage() {
-  redirect('/story');
-}
