--- v2/src/app/pitch/deck/page.tsx
@@ -1,11 +1,0 @@
-import { DeckPublic } from './deck-public';
-
-export const metadata = {
-  title: 'The Goods story deck | Goods on Country',
-  description:
-    'Ten turns through what communities have built with Goods: the need, the products they designed, the making moving into community hands, and what moves next. Real voices, real photographs, audited numbers.',
-};
-
-export default function PitchDeckPage() {
-  return <DeckPublic />;
-}
