+++ v2/src/app/story/page.tsx
@@ -1,1178 +1,327 @@
-import Link from 'next/link';
-import Image from 'next/image';
-import { brand, story, quotes, communityPartnerships, videoTestimonials, partners, advisoryGroup } from '@/lib/data/content';
-import { media, videoUrl } from '@/lib/data/media';
-import { empathyLedger } from '@/lib/empathy-ledger';
-import { MediaSlot, VideoSlot } from '@/components/ui/media-slot';
-import { Button } from '@/components/ui/button';
-import { CANONICAL_ASSETS } from '@/lib/data/asset-canonical';
-import { PLASTIC_KG_PER_BED, STRETCH_BED } from '@/lib/data/products';
+/**
+ * /story/road — the Goods story at long length, on the road spine.
+ *
+ * Content lives in `lib/data/story-road.ts`, never in this file. The current
+ * `/story` page hardcodes 1,178 lines of prose into JSX, which is why its copy
+ * drifted out of reach of every prose guard in the repo. Markup here, words there.
+ *
+ * Ships alongside `/story` rather than replacing it, so both can be walked before
+ * anything is switched. Replacing `/story` is a redirect and a decision, not a
+ * rewrite.
+ *
+ * VOICES resolve through the storyteller registry at render time, and a voice whose
+ * tier is not `external` is DROPPED rather than rendered. The registry is the gate;
+ * this page trusts it and does not keep its own list.
+ *
+ * GAPS render. A stop that is short a photo says so, in place. Silent fallback is how
+ * you stop knowing what you are missing.
+ */
+
 import type { Metadata } from 'next';
+import Image from 'next/image';
+import Link from 'next/link';
+import {
+  storyStops,
+  storyOpening,
+  storyUpdated,
+  type StoryStop,
+  type StoryFigure,
+  type StoryImage,
+  type StoryGap,
+  type StoryVideo,
+} from '@/lib/data/story-road';
+import { getStoryteller, type StorytellerRecord } from '@/lib/data/storyteller-registry';
 
 export const metadata: Metadata = {
-  title: 'Our Story',
-  description: 'How a question about preventable disease became a movement for community ownership. The Goods on Country story, from floor to bed, from charity to commerce.',
+  title: 'The Work That Stays',
+  description:
+    'The full Goods on Country history, walked as a road: Kalgoorlie, Tennant Creek, Palm Island, Utopia, the farm and Maningrida, and Alice Springs. The model arrives at the end, because it is what the road produced.',
 };
 
-const normanQuote = quotes.find(q => q.author === 'Norman Frank' && q.theme === 'community-need');
-const codesignQuote = quotes.find(q => q.author === 'Dianne Stokes');
+// ─────────────────────────────────────────────────────────── voices
 
-// ─── Video slots ───────────────────────────────────────────
-// Cut from raw footage in /media/raw/
-// Hero + Stretch Bed + Building Together from "Stretch Bed Overlay.mp4"
-// Jaquilane testimony from "Jaquilane mum full.mp4"
-const STORY_VIDEOS = {
-  // Hero: community people arriving + assembling beds (0:20-1:20 from Stretch Bed Overlay)
-  heroCommunity: {
-    desktop: videoUrl('community-desktop.mp4'),
-    mobile: videoUrl('community-mobile.mp4'),
-    poster: '/video/community-poster.jpg',
-  },
-  // Stretch Bed: close-up assembly/weave detail (1:10-2:30 from Stretch Bed Overlay)
-  stretchBed: {
-    desktop: videoUrl('stretch-bed-desktop.mp4'),
-    mobile: videoUrl('stretch-bed-mobile.mp4'),
-    poster: '/video/stretch-bed-poster.jpg',
-  },
-  // Young person helping build: group building (0:35-1:10 from Stretch Bed Overlay)
-  buildingTogether: {
-    desktop: videoUrl('building-together-desktop.mp4'),
-    mobile: videoUrl('building-together-mobile.mp4'),
-    poster: '/video/building-together-poster.jpg',
-  },
-  // Jaquilane: overlay background (silent, 0:05-0:35 from Jaquilane mum full)
-  jaquilaneOverlay: {
-    desktop: videoUrl('jaquilane-overlay-desktop.mp4'),
-    mobile: videoUrl('jaquilane-overlay-mobile.mp4'),
-    poster: '/video/jaquilane-poster.jpg',
-  },
-  // Jaquilane: full testimony with audio (0:10-2:10 from Jaquilane mum full)
-  jaquilaneTestimony: videoUrl('jaquilane-testimony.mp4'),
-  washingMachine: undefined as string | undefined,      // Replace with washing machine background video
-  projectStats: undefined as string | undefined,         // Full-page stats overlay video
-  dianneStokes: undefined as string | undefined,         // Dianne Stokes washing machine video (Descript)
-  // Recycling plant: containerised factory footage (0:00-0:50 from recycling gear.mp4)
-  recyclingPlant: {
-    desktop: videoUrl('recycling-plant-desktop.mp4'),
-    mobile: videoUrl('recycling-plant-mobile.mp4'),
-    poster: '/video/recycling-plant-poster.jpg',
-  },
-  cliffPlummer: videoTestimonials[0]?.embedUrl,           // Cliff Plummer: beds and dignity
-};
+/**
+ * Resolve a name to a renderable voice, or null.
+ *
+ * Default-deny in two places: an unknown name resolves to tier `hold` in the
+ * registry and is dropped here, and a person with no usable quote is dropped too.
+ * `hold` and `retired` quotes can never be selected.
+ */
+function resolveVoice(name: string): { person: StorytellerRecord; quote: string; context: string } | null {
+  const person = getStoryteller(name);
+  if (!person || person.tier !== 'external') return null;
 
-// ─── Photo gallery slots ───────────────────────────────────
-// These define the initial photo sets for product detail sections.
-// The "Add photo" button in each section links to an admin upload flow.
-const STRETCH_BED_PHOTOS = [
-  { src: media.product.stretchBedHero, alt: 'Stretch Bed alone at golden hour showing recycled plastic X-legs', label: 'Product shot' },
-  { src: media.product.stretchBedInUse, alt: 'Person resting on Stretch Bed in the outback', label: 'In use' },
-  { src: media.product.stretchBedAssembly, alt: 'Two people assembling a Stretch Bed together', label: 'Assembly' },
-  { src: media.product.stretchBedCommunity, alt: 'Elder woman presenting her new Stretch Bed', label: 'Community' },
-];
+  const usable = person.quotes.filter((q) => q.status === 'primary' || q.status === 'approved');
+  const chosen = usable.find((q) => q.status === 'primary') ?? usable[0];
+  if (!chosen) return null;
 
-const WASHING_MACHINE_PHOTOS = [
-  { src: media.product.washingMachine, alt: 'Pakkimjalki Kari washing machine with recycled plastic enclosure', label: 'Product shot' },
-  { src: media.product.washingMachineName, alt: 'Close-up showing Pakkimjalki Kari name in Warumungu', label: 'Name detail' },
-  { src: media.product.washingMachineInstalled, alt: 'Washing machine installed in community laundry', label: 'Installed' },
-];
+  return { person, quote: chosen.text, context: chosen.context };
+}
 
-const RECYCLING_PLANT_PHOTOS = [
-  { src: media.manufacturing.containerFactory, alt: 'Containerised recycling factory with shredder', label: 'Container factory' },
-  { src: media.manufacturing.hydraulicPress, alt: 'Hydraulic press forming recycled plastic sheets', label: 'Press' },
-  { src: media.manufacturing.pressedSheets, alt: 'Stack of colorful recycled plastic sheets', label: 'Pressed sheets' },
-];
+function Voice({ name }: { name: string }) {
+  const voice = resolveVoice(name);
+  if (!voice) return null;
+  const { person, quote, context } = voice;
 
-const TIMELINE = [
-  { year: '2018', title: 'The Spark', description: 'Nic hears Dr. Bo Remenyi speak about Rheumatic Heart Disease. Entirely preventable, yet children in remote communities are dying from it.' },
-  { year: '2020', title: 'The Pattern', description: 'Orange Sky expands to remote communities. Nic sees people without beds, without washing machines, children with skin infections cascading into heart disease.' },
-  { year: '2022', title: 'The Question', description: '"What can we actually build that makes a difference?" Goods project kicks off with an advisory session in November.' },
-  { year: '2023', title: 'A Curious Tractor', description: 'Organisation formally founded. First bed prototypes developed with the Bloomfield family "around the fire."' },
-  { year: '2024', title: '540 Beds', description: 'Active pilots deliver beds across Palm Island, Tennant Creek, Mt Isa, Kalgoorlie, and more. Community feedback shapes every iteration.' },
-  { year: '2025', title: 'Washing Machines', description: 'Pakkimjalki Kari, named in Warumungu language by Elder Dianne Stokes. Commercial-grade, one-button operation. 11 communities served.' },
-];
-
-function PhotoGallery({ photos }: { photos: { src: string | undefined; alt: string; label: string }[] }) {
   return (
-    <div className="grid grid-cols-2 gap-3">
-      {photos.map((photo, i) => (
-        <MediaSlot
-          key={i}
-          src={photo.src}
-          alt={photo.alt}
-          label={photo.label}
-          aspect="4/3"
-        />
+    <figure className="my-10 border-l-2 border-[var(--goods-terracotta,#B7593F)] pl-6 md:pl-8">
+      <blockquote className="font-serif text-2xl md:text-3xl leading-snug text-foreground">
+        {quote}
+      </blockquote>
+      <figcaption className="mt-4 flex items-center gap-3 text-sm">
+        {person.portrait && (
+          <Image
+            src={person.portrait}
+            alt=""
+            width={44}
+            height={44}
+            className="rounded-full object-cover h-11 w-11"
+          />
+        )}
+        <span>
+          <span className="font-medium text-foreground">{person.name}</span>
+          <span className="block text-muted-foreground">
+            {/* Practitioners must be labelled as practitioners, never as community
+                recipients. The registry carries the flag; the UI has to honour it. */}
+            {person.practitioner ? `${person.role} · practitioner` : person.role}
+            {person.community ? ` · ${person.community}` : ''}
+          </span>
+          <span className="block text-muted-foreground/70 text-xs mt-0.5">{context}</span>
+        </span>
+      </figcaption>
+    </figure>
+  );
+}
+
+// ─────────────────────────────────────────────────────────── parts
+
+function Figures({ figures }: { figures: StoryFigure[] }) {
+  return (
+    <dl className="my-10 grid gap-x-8 gap-y-6 grid-cols-2 md:grid-cols-4">
+      {figures.map((f) => (
+        <div key={f.label}>
+          <dt className="font-serif text-3xl md:text-4xl text-foreground">{f.value}</dt>
+          <dd className="mt-1 text-sm text-muted-foreground">{f.label}</dd>
+          {/* Claims status is not decoration. A modelled figure and a verified one
+              must never look the same on a funder-facing page. */}
+          <dd className="mt-1.5 inline-block rounded-full border px-2 py-0.5 text-[11px] uppercase tracking-wide text-muted-foreground">
+            {f.claim}
+          </dd>
+        </div>
+      ))}
+    </dl>
+  );
+}
+
+/**
+ * Local clips are click-to-play behind their poster: `preload="none"` means the
+ * file is not fetched until somebody presses play. On a page with five clips that
+ * is the difference between a poster and forty megabytes, which matters most for
+ * exactly the people this story is about.
+ */
+function Video({ video }: { video: StoryVideo }) {
+  return (
+    <figure className="my-10">
+      <div className="relative w-full overflow-hidden rounded-md aspect-video bg-muted">
+        {video.kind === 'local' ? (
+          <video
+            controls
+            preload="none"
+            poster={video.poster}
+            className="absolute inset-0 h-full w-full object-cover"
+          >
+            <source src={video.src} type="video/mp4" />
+          </video>
+        ) : (
+          <iframe
+            src={video.embedUrl}
+            title={video.label}
+            allowFullScreen
+            loading="lazy"
+            className="absolute inset-0 h-full w-full border-0"
+          />
+        )}
+      </div>
+      <figcaption className="mt-2 text-sm text-muted-foreground">
+        {video.label}
+        {video.caption && <span className="block mt-0.5">{video.caption}</span>}
+      </figcaption>
+    </figure>
+  );
+}
+
+/**
+ * Gap markers are a REVIEW tool, not public copy.
+ *
+ * They exist so a missing photo or voice stays countable instead of being papered
+ * over by a fallback. But they read as unfinished to a funder, and their text
+ * discusses editorial state, so they render only for someone who asked to see
+ * them: in development, or on ?review=1.
+ *
+ * This is not the same as hiding them. `storyGaps()` still enumerates every one,
+ * and the tests still hold them. The gap stays counted; it just is not published.
+ */
+function Gap({ gap }: { gap: StoryGap }) {
+  return (
+    <div className="my-8 rounded-md border border-dashed px-5 py-4 text-sm text-muted-foreground">
+      <span className="font-medium text-foreground">
+        Missing {gap.slot}: {gap.wanted}
+      </span>
+      <span className="block mt-1">{gap.reason}</span>
+    </div>
+  );
+}
+
+function Gallery({ images }: { images: StoryImage[] }) {
+  return (
+    <div className={`my-10 grid gap-4 ${images.length > 1 ? 'sm:grid-cols-2' : ''}`}>
+      {images.map((img) => (
+        <figure key={img.src}>
+          <Image
+            src={img.src}
+            alt={img.alt}
+            width={900}
+            height={600}
+            sizes="(max-width: 640px) 100vw, 45vw"
+            className="w-full h-auto rounded-md object-cover"
+          />
+          {img.caption && (
+            <figcaption className="mt-2 text-sm text-muted-foreground">{img.caption}</figcaption>
+          )}
+        </figure>
       ))}
     </div>
   );
 }
 
-export default async function StoryPage() {
-  // Fetch storyteller profiles from Empathy Ledger for avatar images
-  // Fetch all storytellers (239 total) to ensure we match all quote authors
-  const storytellers = await empathyLedger.getStorytellers({ limit: 250 });
-  const avatarMap: Record<string, string> = {};
-  for (const s of storytellers) {
-    if (s.avatarUrl) {
-      // Normalize name: trim whitespace, collapse multiple spaces
-      const normalized = s.displayName.trim().replace(/\s+/g, ' ');
-      avatarMap[normalized] = s.avatarUrl;
-    }
-  }
+function Stop({ stop, showGaps }: { stop: StoryStop; showGaps: boolean }) {
   return (
-    <main className="scroll-smooth scroll-pt-12">
-      {/* ─── Sticky section nav ─── */}
-      <nav className="sticky top-0 z-50 bg-background/80 backdrop-blur-md border-b border-border/50">
-        <div className="container mx-auto px-4">
-          <div className="flex items-center gap-1 overflow-x-auto py-2 text-xs no-scrollbar">
-            {[
-              { id: 'hero', label: 'Home' },
-              { id: 'problem', label: 'The Problem' },
-              { id: 'cascade', label: 'The Cascade' },
-              { id: 'stretch-bed', label: 'Stretch Bed' },
-              { id: 'washing-machine', label: 'Washing Machine' },
-              { id: 'journey', label: 'Journey' },
-              { id: 'voices', label: 'Voices' },
-              { id: 'manufacturing', label: 'Manufacturing' },
-              { id: 'impact', label: 'Impact' },
-              { id: 'partners', label: 'Partners' },
-              { id: 'future', label: 'What\'s Next' },
-            ].map((item) => (
-              <a
-                key={item.id}
-                href={`#${item.id}`}
-                className="shrink-0 px-3 py-1.5 rounded-full hover:bg-muted text-muted-foreground hover:text-foreground transition-colors whitespace-nowrap"
-              >
-                {item.label}
-              </a>
-            ))}
-          </div>
-        </div>
-      </nav>
+    <section id={stop.id} className="scroll-mt-24 border-t py-16 md:py-24">
+      <p className="text-sm uppercase tracking-[0.14em] text-muted-foreground">{stop.eyebrow}</p>
+      <h2 className="mt-3 font-serif text-4xl md:text-5xl leading-tight text-foreground">
+        {stop.headline}
+      </h2>
+      {stop.place && <p className="mt-3 text-sm text-muted-foreground">{stop.place}</p>}
 
-      <div className="story-scroll">
-
-        {/* ════════════════════════════════════════════════════════════════
-            1. HERO: Community building video (NOT timelapse)
-            ════════════════════════════════════════════════════════════ */}
-        <section id="hero" className="min-h-screen flex items-center justify-center relative overflow-hidden">
-          {/* Background video: community people building */}
-          <video
-            className="absolute inset-0 w-full h-full object-cover"
-            autoPlay
-            muted
-            loop
-            playsInline
-            poster={STORY_VIDEOS.heroCommunity.poster}
-          >
-            <source src={STORY_VIDEOS.heroCommunity.desktop} media="(min-width: 768px)" type="video/mp4" />
-            <source src={STORY_VIDEOS.heroCommunity.mobile} type="video/mp4" />
-          </video>
-          <div className="absolute inset-0 bg-gradient-to-t from-black/80 via-black/50 to-black/30" />
-
-          <div className="relative z-10 container mx-auto px-4 text-center">
-            <p className="text-sm uppercase tracking-widest text-primary mb-6">
-              {brand.oneLiner}
-            </p>
-            <h1
-              className="text-4xl md:text-6xl lg:text-7xl font-light text-white mb-6 leading-[1.1] max-w-4xl mx-auto"
-              style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-            >
-              Built with communities, not for them.
-            </h1>
-            <p className="text-xl text-white/60 max-w-2xl mx-auto">
-              Beds, washing machines, and essential goods designed with remote Indigenous communities. Manufactured sustainably, eventually owned by them.
-            </p>
-          </div>
-
-          <div className="absolute bottom-8 left-1/2 -translate-x-1/2 z-10 animate-bounce">
-            <svg className="w-5 h-5 text-white/30" fill="none" viewBox="0 0 24 24" strokeWidth="2" stroke="currentColor">
-              <path strokeLinecap="round" strokeLinejoin="round" d="M19.5 13.5L12 21m0 0l-7.5-7.5M12 21V3" />
-            </svg>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            2. THE PROBLEM: Stats
-            ════════════════════════════════════════════════════════════ */}
-        <section id="problem" className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto text-center">
-              <p className="text-sm uppercase tracking-widest text-primary mb-4">The Problem</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-foreground mb-10 leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                A failure of infrastructure, not culture
-              </h2>
-              <div className="grid grid-cols-2 md:grid-cols-4 gap-6 mb-12">
-                {story.problem.stats.map((stat) => (
-                  <div key={stat.label} className="text-center">
-                    <p className="text-3xl md:text-4xl font-bold text-primary">{stat.value}</p>
-                    <p className="text-muted-foreground text-sm mt-1">{stat.label}</p>
-                  </div>
-                ))}
-              </div>
-              <p className="text-lg text-muted-foreground leading-relaxed max-w-2xl mx-auto">
-                {story.problem.description}
-              </p>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            2b. THE CASCADE: Health intervention
-            ════════════════════════════════════════════════════════════ */}
-        <section id="cascade" className="min-h-screen flex items-center justify-center bg-foreground text-background">
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-background/40 mb-4 text-center">The Cascade</p>
-              <h2
-                className="text-3xl md:text-4xl font-light text-center mb-12 leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                Why a washing machine is<br />a health intervention
-              </h2>
-              <div className="flex flex-col items-center gap-0">
-                {[
-                  { text: 'No washing machine', stat: '59% of remote homes' },
-                  { text: 'Dirty bedding', stat: null },
-                  { text: 'Scabies', stat: '1 in 3 children affected' },
-                  { text: 'Skin infections → Strep A', stat: null },
-                  { text: 'Rheumatic fever', stat: null },
-                  { text: 'Rheumatic Heart Disease', stat: 'entirely preventable' },
-                ].map((step, i, arr) => (
-                  <div key={i} className="flex flex-col items-center">
-                    <div className={`px-6 py-3 rounded-xl text-center ${i === arr.length - 1 ? 'bg-primary/20 border border-primary/40' : 'bg-white/5 border border-white/10'}`}>
-                      <span className={`text-lg font-medium ${i === arr.length - 1 ? 'text-primary' : 'text-white'}`}>{step.text}</span>
-                      {step.stat && (
-                        <span className="block text-xs text-white/40 mt-0.5">{step.stat}</span>
-                      )}
-                    </div>
-                    {i < arr.length - 1 && (
-                      <svg className="w-4 h-6 text-white/20 my-1" fill="none" viewBox="0 0 16 24" strokeWidth="2" stroke="currentColor">
-                        <path strokeLinecap="round" strokeLinejoin="round" d="M8 4v16m0 0l-4-4m4 4l4-4" />
-                      </svg>
-                    )}
-                  </div>
-                ))}
-              </div>
-              <div className="mt-12 text-center">
-                <p
-                  className="text-xl md:text-2xl font-light text-primary leading-relaxed"
-                  style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                >
-                  A washing machine. Clean bedding. A dry, washable bed.<br />
-                  This is health hardware, not furniture.<br />
-                  We build the conditions for healthier homes.
-                </p>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            3. THE VOICE: Norman Frank, full-page image background
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center relative overflow-hidden">
-          {/* Full-page Norman Frank portrait */}
+      {stop.photo && (
+        <figure className="my-10">
           <Image
-            src={avatarMap[normanQuote?.author ?? ''] || media.people.normFrank || '/images/people/norman-frank.jpg'}
-            alt={normanQuote?.author ?? 'Norman Frank'}
-            fill
-            className="object-cover"
-            sizes="100vw"
-            priority
+            src={stop.photo.src}
+            alt={stop.photo.alt}
+            width={1400}
+            height={900}
+            sizes="(max-width: 768px) 100vw, 70vw"
+            className="w-full h-auto rounded-md object-cover"
           />
-          <div className="absolute inset-0 bg-gradient-to-t from-black/80 via-black/50 to-black/30" />
-          <div className="relative z-10 container mx-auto px-4">
-            <div className="max-w-3xl mx-auto text-center">
-              <blockquote>
-                <p
-                  className="text-3xl md:text-4xl lg:text-5xl font-light text-white leading-snug mb-8 drop-shadow-lg"
-                  style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                >
-                  &ldquo;{normanQuote?.text}&rdquo;
-                </p>
-                <footer>
-                  <span className="font-medium text-white text-lg drop-shadow-md">{normanQuote?.author}</span>
-                  <br />
-                  <span className="text-white/70 text-sm">{normanQuote?.context}</span>
-                </footer>
-              </blockquote>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            4. THE ORIGIN: How it started
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-primary mb-4">The Origin</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-foreground mb-8 leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                How a question became a movement
-              </h2>
-              <div className="space-y-6 text-lg text-muted-foreground leading-relaxed">
-                <p>
-                  In 2018, Nicholas Marchesi attended a health conference where Dr. Bo Remenyi spoke about
-                  Rheumatic Heart Disease. Her message was clear: RHD is entirely preventable, yet she was
-                  filling out death certificates for children in remote communities.
-                </p>
-                <p>
-                  Combined with his experience at Orange Sky, where he witnessed people without beds,
-                  without working washing machines, and children with skin infections that cascade into
-                  heart disease, it sparked the question:
-                </p>
-                <p
-                  className="text-2xl md:text-3xl font-light text-foreground"
-                  style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                >
-                  What can we actually build that makes a difference?
-                </p>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            5. STRETCH BED: Full-page video with overlay
-            ════════════════════════════════════════════════════════════ */}
-        <section id="stretch-bed" className="min-h-screen flex items-center justify-center relative overflow-hidden">
-          <video
-            className="absolute inset-0 w-full h-full object-cover"
-            autoPlay
-            muted
-            loop
-            playsInline
-            poster={STORY_VIDEOS.stretchBed.poster}
-          >
-            <source src={STORY_VIDEOS.stretchBed.desktop} media="(min-width: 768px)" type="video/mp4" />
-            <source src={STORY_VIDEOS.stretchBed.mobile} type="video/mp4" />
-          </video>
-          <div className="absolute inset-0 bg-black/50" />
-          <div className="relative z-10 container mx-auto px-4 text-center">
-            <p className="text-sm uppercase tracking-widest text-primary mb-6">The Stretch Bed</p>
-            <h2
-              className="text-4xl md:text-6xl font-light text-white mb-6 leading-[1.1]"
-              style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-            >
-              {STRETCH_BED.specs.weight}. Supports {STRETCH_BED.specs.loadCapacity}.<br />
-              No tools. 5 minutes.
-            </h2>
-            <p className="text-xl text-white/60 max-w-2xl mx-auto">
-              Recycled plastic legs, galvanised steel poles, heavy-duty canvas.
-              Each bed diverts {PLASTIC_KG_PER_BED}kg of plastic from landfill.
-            </p>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            5b. STRETCH BED: Detail + photos
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-5xl mx-auto">
-              <div className="grid gap-12 lg:grid-cols-2 items-center">
-                <div>
-                  <p className="text-sm uppercase tracking-widest text-primary mb-3">Product Detail</p>
-                  <h2
-                    className="text-3xl md:text-4xl font-light text-foreground mb-6 leading-snug"
-                    style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                  >
-                    The Stretch Bed
-                  </h2>
-                  <div className="space-y-4 text-muted-foreground leading-relaxed mb-8">
-                    <p>
-                      Two galvanised steel poles thread through heavy-duty Australian canvas.
-                      Two X-trestle legs pressed from recycled HDPE plastic, with production designed to move toward On-Country.
-                      Each bed diverts {PLASTIC_KG_PER_BED}kg of plastic from landfill.
-                    </p>
-                    <p>
-                      No-tool assembly in 5 minutes. Works inside and outside. Fully washable.
-                      Designed with communities, not for them.
-                    </p>
-                  </div>
-                  <div className="grid grid-cols-2 gap-3 mb-6">
-                    <div className="text-center p-4 rounded-xl border border-border bg-background">
-                      <div className="text-2xl font-bold text-primary">$750</div>
-                      <div className="text-xs text-muted-foreground mt-1">Stretch Bed</div>
-                    </div>
-                    <div className="text-center p-4 rounded-xl border border-border bg-background">
-                      <div className="text-2xl font-bold text-primary">$1,500</div>
-                      <div className="text-xs text-muted-foreground mt-1">avg. bed &amp; base in community</div>
-                    </div>
-                  </div>
-                  <div className="flex gap-3">
-                    <Button asChild>
-                      <Link href="/shop/stretch-bed-single">Shop Stretch Bed</Link>
-                    </Button>
-                    <Button variant="outline" asChild>
-                      <Link href="/process">How It&rsquo;s Made</Link>
-                    </Button>
-                  </div>
-                </div>
-                <div>
-                  <PhotoGallery photos={STRETCH_BED_PHOTOS} />
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            5c/5d. JAQUILANE testimony PULLED 2026-07-03 — consent clearance is
-            in conflict across our records, so both the overlay and the embedded
-            testimony are removed from the live page until a single source of
-            truth confirms clearance. Assets remain in the repo (jaquilane-*.mp4)
-            and STORY_VIDEOS.jaquilane*; re-add this section once cleared.
-            ════════════════════════════════════════════════════════════ */}
-
-        {/* ════════════════════════════════════════════════════════════════
-            6. WASHING MACHINE: Full-page video with overlay
-            ════════════════════════════════════════════════════════════ */}
-        <section id="washing-machine" className="min-h-screen flex items-center justify-center relative overflow-hidden bg-foreground">
-          {/* Background: video if available, otherwise product image or gradient */}
-          {STORY_VIDEOS.washingMachine ? (
-            <video
-              className="absolute inset-0 w-full h-full object-cover"
-              autoPlay
-              muted
-              loop
-              playsInline
-            >
-              <source src={STORY_VIDEOS.washingMachine} type="video/mp4" />
-            </video>
-          ) : media.product.washingMachine ? (
-            <Image
-              src={media.product.washingMachine}
-              alt="Pakkimjalki Kari washing machine"
-              fill
-              className="object-cover"
-              sizes="100vw"
-            />
-          ) : (
-            <div className="absolute inset-0 bg-gradient-to-br from-foreground via-foreground to-primary/20" />
+          {stop.photo.caption && (
+            <figcaption className="mt-2 text-sm text-muted-foreground">
+              {stop.photo.caption}
+            </figcaption>
           )}
-          <div className="absolute inset-0 bg-black/40" />
-          <div className="relative z-10 container mx-auto px-4 text-center">
-            <p className="text-sm uppercase tracking-widest text-primary mb-6">Pakkimjalki Kari</p>
-            <h2
-              className="text-4xl md:text-6xl font-light text-white mb-6 leading-[1.1]"
-              style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-            >
-              A washing machine isn&rsquo;t convenience.<br />
-              It&rsquo;s cardiac prevention.
-            </h2>
-            <p className="text-xl text-white/60 max-w-2xl mx-auto">
-              Named in Warumungu language by Elder Dianne Stokes.
-              Commercial-grade Speed Queen base. One-button operation. Built for remote conditions.
-            </p>
-          </div>
-        </section>
+        </figure>
+      )}
 
-        {/* ════════════════════════════════════════════════════════════════
-            6b. WASHING MACHINE: Detail + photos
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center bg-background">
-          <div className="container mx-auto px-4">
-            <div className="max-w-5xl mx-auto">
-              <div className="grid gap-12 lg:grid-cols-2 items-center">
-                <div>
-                  <p className="text-sm uppercase tracking-widest text-primary mb-3">Product Detail</p>
-                  <h2
-                    className="text-3xl md:text-4xl font-light text-foreground mb-6 leading-snug"
-                    style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                  >
-                    Pakkimjalki Kari
-                  </h2>
-                  <div className="space-y-4 text-muted-foreground leading-relaxed mb-8">
-                    <p>
-                      Commercial-grade Speed Queen base modified for remote conditions.
-                      One-button operation, no complex cycles. Built to last 10-15 years, not 1-2.
-                    </p>
-                    <p>
-                      Clean bedding breaks the scabies cycle. Scabies leads to Rheumatic Heart Disease.
-                      The washing machine is a cardiac prevention tool.
-                    </p>
-                    <p className="text-sm">
-                      One Alice Springs provider sells $3M of washing machines annually into communities.
-                      Most end up in dumps within 12 months. These communities deserve better.
-                    </p>
-                  </div>
-                  <Button asChild>
-                    <Link href="/partner">Register Interest</Link>
-                  </Button>
-                </div>
-                <div>
-                  <PhotoGallery photos={WASHING_MACHINE_PHOTOS} />
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            7. THE JOURNEY: Timeline
-            ════════════════════════════════════════════════════════════ */}
-        <section id="journey" className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-4xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-primary mb-4 text-center">The Journey</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-foreground mb-12 text-center leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                From question to community
-              </h2>
-              <div className="grid gap-4 md:grid-cols-2 lg:grid-cols-3">
-                {TIMELINE.map((item) => (
-                  <div key={item.year} className="rounded-xl border border-border p-6 bg-background">
-                    <div className="text-2xl font-bold text-primary mb-1">{item.year}</div>
-                    <h3 className="font-semibold text-foreground mb-2">{item.title}</h3>
-                    <p className="text-sm text-muted-foreground leading-relaxed">{item.description}</p>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            8. COMMUNITY VOICES: Full-bleed quote panels
-            ════════════════════════════════════════════════════════════ */}
-        <div id="voices" />
-        {[
-          { quote: quotes.find(q => q.author === 'Alfred Johnson' && q.theme === 'freight-tax'), fallbackImage: media.community.palmIsland || '/images/community/palm-island.jpg' },
-          { quote: quotes.find(q => q.author === 'Chloe'), fallbackImage: '/images/community/tennant-creek.jpg' },
-          { quote: quotes.find(q => q.author === 'Patricia Frank'), fallbackImage: '/images/community/tennant-creek.jpg' },
-          { quote: quotes.find(q => q.author === 'Linda Turner'), fallbackImage: media.community.tennantCreek || '/images/community/tennant-creek.jpg' },
-        ].filter(panel => panel.quote).map((panel, i) => (
-          <section key={`voice-${i}`} className="min-h-screen flex items-center justify-center relative overflow-hidden">
-            <Image
-              src={avatarMap[panel.quote!.author] || panel.fallbackImage}
-              alt={panel.quote!.author}
-              fill
-              className="object-cover"
-              sizes="100vw"
-            />
-            <div className="absolute inset-0 bg-gradient-to-t from-black/85 via-black/60 to-black/40" />
-            <div className="relative z-10 container mx-auto px-4">
-              <div className="max-w-3xl mx-auto text-center">
-                <blockquote>
-                  <p
-                    className="text-2xl md:text-4xl lg:text-5xl font-light text-white leading-snug mb-8 drop-shadow-lg"
-                    style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                  >
-                    &ldquo;{panel.quote!.text}&rdquo;
-                  </p>
-                  <footer className="flex items-center justify-center gap-3">
-                    {avatarMap[panel.quote!.author] ? (
-                      <Image
-                        src={avatarMap[panel.quote!.author]}
-                        alt={panel.quote!.author}
-                        width={48}
-                        height={48}
-                        className="rounded-full object-cover w-12 h-12 border-2 border-white/30"
-                      />
-                    ) : (
-                      <div className="w-12 h-12 rounded-full bg-white/10 flex items-center justify-center text-white text-lg font-medium border-2 border-white/30">
-                        {panel.quote!.author.charAt(0)}
-                      </div>
-                    )}
-                    <div className="text-left">
-                      <span className="font-medium text-white text-lg drop-shadow-md">{panel.quote!.author}</span>
-                      <br />
-                      <span className="text-white/60 text-sm">{panel.quote!.context}</span>
-                    </div>
-                  </footer>
-                </blockquote>
-              </div>
-            </div>
-          </section>
+      <div className="mt-8 space-y-6">
+        {stop.chapters.map((para, i) => (
+          <p key={i} className="text-lg md:text-xl leading-relaxed text-foreground/90">
+            {para}
+          </p>
         ))}
-
-        {/* Discrete entry point to the full voices grid */}
-        <section className="py-12 bg-foreground text-background">
-          <div className="container mx-auto px-4 text-center">
-            <Link
-              href="/stories"
-              className="inline-flex items-center gap-2 text-sm uppercase tracking-[0.25em] text-background/70 hover:text-background transition-colors"
-            >
-              Read all community voices
-              <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" strokeWidth="1.5" stroke="currentColor">
-                <path strokeLinecap="round" strokeLinejoin="round" d="M13.5 4.5L21 12m0 0l-7.5 7.5M21 12H3" />
-              </svg>
-            </Link>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            8b. FREIGHT TAX: The cost of remoteness
-            ════════════════════════════════════════════════════════════ */}
-        <section className="flex items-center justify-center py-16 bg-foreground">
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto text-center">
-              <p
-                className="text-2xl md:text-3xl font-light text-white/80 leading-relaxed"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                <span className="text-primary text-4xl md:text-5xl font-bold">$1,200+</span>
-                <br />
-                for a mattress in remote areas, 2&times; the city price.
-                <br />
-                <span className="text-white/40 text-lg">Lasts weeks, not years.</span>
-              </p>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            8c. BUILDING TOGETHER: Young person helping build
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-end relative overflow-hidden">
-          <video
-            className="absolute inset-0 w-full h-full object-cover"
-            autoPlay
-            muted
-            loop
-            playsInline
-            poster={STORY_VIDEOS.buildingTogether.poster}
-          >
-            <source src={STORY_VIDEOS.buildingTogether.desktop} media="(min-width: 768px)" type="video/mp4" />
-            <source src={STORY_VIDEOS.buildingTogether.mobile} type="video/mp4" />
-          </video>
-          <div className="absolute inset-0 bg-gradient-to-t from-black/70 via-black/20 to-transparent" />
-          <div className="relative z-10 w-full pb-16 md:pb-20">
-            <div className="container mx-auto px-4">
-              <div className="max-w-2xl">
-                <p className="text-sm uppercase tracking-widest text-primary mb-3">Building Together</p>
-                <h2
-                  className="text-3xl md:text-4xl font-light text-white mb-4 leading-snug"
-                  style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                >
-                  Beds aren&rsquo;t delivered.<br />
-                  They&rsquo;re assembled together.
-                </h2>
-                <p className="text-lg text-white/60 max-w-lg">
-                  Every delivery is an event. Families gather, young people help assemble,
-                  and community grows stronger, one bed at a time.
-                </p>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            8c. RECYCLING PLANT: Full-page video overlay
-            ════════════════════════════════════════════════════════════ */}
-        <section id="manufacturing" className="min-h-screen flex items-center justify-center relative overflow-hidden">
-          <video
-            className="absolute inset-0 w-full h-full object-cover"
-            autoPlay
-            muted
-            loop
-            playsInline
-            poster={STORY_VIDEOS.recyclingPlant.poster}
-          >
-            <source src={STORY_VIDEOS.recyclingPlant.desktop} media="(min-width: 768px)" type="video/mp4" />
-            <source src={STORY_VIDEOS.recyclingPlant.mobile} type="video/mp4" />
-          </video>
-          <div className="absolute inset-0 bg-black/55" />
-          <div className="relative z-10 container mx-auto px-4 text-center">
-            <p className="text-sm uppercase tracking-widest text-primary mb-6">The Recycling Plant</p>
-            <h2
-              className="text-4xl md:text-6xl font-light text-white mb-6 leading-[1.1]"
-              style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-            >
-              Local waste into<br />
-              local products.
-            </h2>
-            <p className="text-xl text-white/60 max-w-2xl mx-auto">
-              A containerised factory that travels to communities. Shred, press, cut.
-              Designed to turn plastic waste into bed components, moving toward On-Country production.
-            </p>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            8d. RECYCLING PLANT: Detail + photos
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-5xl mx-auto">
-              <div className="grid gap-12 lg:grid-cols-2 items-center">
-                <div>
-                  <p className="text-sm uppercase tracking-widest text-primary mb-3">On-Country Manufacturing</p>
-                  <h2
-                    className="text-3xl md:text-4xl font-light text-foreground mb-6 leading-snug"
-                    style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                  >
-                    The Portable Factory
-                  </h2>
-                  <div className="space-y-4 text-muted-foreground leading-relaxed mb-8">
-                    <p>
-                      A fully containerised production plant. Fits in a shipping container, sets up in a day.
-                      Designed so a shredder, hydraulic press, and CNC cutter can turn community plastic waste
-                      into precision-cut bed components, moving toward On-Country production.
-                    </p>
-                    <p>
-                      Capacity: ~30 beds per week. Each bed diverts {PLASTIC_KG_PER_BED}kg of HDPE plastic from landfill.
-                      Built so local people can operate the machinery: real jobs, real skills, a pathway to local ownership.
-                    </p>
-                  </div>
-                  <div className="grid grid-cols-3 gap-3 mb-6">
-                    <div className="text-center p-4 rounded-xl border border-border bg-background">
-                      <div className="text-2xl font-bold text-primary">30</div>
-                      <div className="text-xs text-muted-foreground mt-1">beds/week</div>
-                    </div>
-                    <div className="text-center p-4 rounded-xl border border-border bg-background">
-                      <div className="text-2xl font-bold text-primary">{PLASTIC_KG_PER_BED}kg</div>
-                      <div className="text-xs text-muted-foreground mt-1">plastic per bed</div>
-                    </div>
-                    <div className="text-center p-4 rounded-xl border border-border bg-background">
-                      <div className="text-2xl font-bold text-primary">1 day</div>
-                      <div className="text-xs text-muted-foreground mt-1">to set up</div>
-                    </div>
-                  </div>
-                  <div className="space-y-3 text-sm text-muted-foreground">
-                    <div className="flex items-start gap-2">
-                      <svg className="h-4 w-4 text-primary flex-shrink-0 mt-0.5" fill="none" viewBox="0 0 24 24" strokeWidth="2" stroke="currentColor">
-                        <path strokeLinecap="round" strokeLinejoin="round" d="M4.5 12.75l6 6 9-13.5" />
-                      </svg>
-                      <span><strong>Shredder</strong>: breaks down collected HDPE plastic into chips</span>
-                    </div>
-                    <div className="flex items-start gap-2">
-                      <svg className="h-4 w-4 text-primary flex-shrink-0 mt-0.5" fill="none" viewBox="0 0 24 24" strokeWidth="2" stroke="currentColor">
-                        <path strokeLinecap="round" strokeLinejoin="round" d="M4.5 12.75l6 6 9-13.5" />
-                      </svg>
-                      <span><strong>Hydraulic press</strong>: melts and compresses chips into durable sheets</span>
-                    </div>
-                    <div className="flex items-start gap-2">
-                      <svg className="h-4 w-4 text-primary flex-shrink-0 mt-0.5" fill="none" viewBox="0 0 24 24" strokeWidth="2" stroke="currentColor">
-                        <path strokeLinecap="round" strokeLinejoin="round" d="M4.5 12.75l6 6 9-13.5" />
-                      </svg>
-                      <span><strong>CNC cutter</strong>: precision-cuts bed frame components from sheets</span>
-                    </div>
-                  </div>
-                </div>
-                <div>
-                  <PhotoGallery photos={RECYCLING_PLANT_PHOTOS} />
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            9. IMPACT VIDEO: Full-page stats overlay
-            ════════════════════════════════════════════════════════════ */}
-        <section id="impact" className="min-h-screen flex items-center justify-center relative overflow-hidden bg-foreground">
-          {STORY_VIDEOS.projectStats ? (
-            <video
-              className="absolute inset-0 w-full h-full object-cover"
-              autoPlay
-              muted
-              loop
-              playsInline
-            >
-              <source src={STORY_VIDEOS.projectStats} type="video/mp4" />
-            </video>
-          ) : (
-            <Image
-              src="/video/building-together-poster.jpg"
-              alt="Community building together"
-              fill
-              className="object-cover"
-              sizes="100vw"
-            />
-          )}
-          <div className="absolute inset-0 bg-black/60" />
-          <div className="relative z-10 container mx-auto px-4">
-            <div className="max-w-5xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-white/40 mb-4 text-center">Impact</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-white mb-12 text-center leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                This isn&rsquo;t a concept. It&rsquo;s working.
-              </h2>
-              <div className="grid gap-4 sm:grid-cols-2 lg:grid-cols-4 mb-12">
-                {[
-                  { value: '540', label: 'bed units deployed' },
-                  { value: String(CANONICAL_ASSETS.communitiesServed), label: 'communities served' },
-                  { value: `${PLASTIC_KG_PER_BED}kg`, label: 'plastic per bed diverted' },
-                ].map((stat) => (
-                  <div key={stat.label} className="text-center p-6 rounded-xl bg-white/5 border border-white/10">
-                    <div className="text-3xl md:text-4xl font-bold text-primary mb-1">{stat.value}</div>
-                    <div className="text-sm text-white/50">{stat.label}</div>
-                  </div>
-                ))}
-              </div>
-              <div className="grid gap-4 sm:grid-cols-3">
-                {communityPartnerships.filter(p => p.bedsDelivered > 0).map((p) => (
-                  <div key={p.id} className="rounded-xl bg-white/5 border border-white/10 p-5">
-                    <div className="flex items-start justify-between mb-2">
-                      <div>
-                        <h3 className="font-semibold text-white">{p.name}</h3>
-                        <p className="text-xs text-white/40">{p.region}</p>
-                      </div>
-                      <div className="text-right">
-                        <div className="text-xl font-bold text-primary">{p.bedsDelivered}</div>
-                        <div className="text-xs text-white/40">beds</div>
-                      </div>
-                    </div>
-                    <p className="text-sm text-white/60">{p.highlight}</p>
-                  </div>
-                ))}
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            10. THE MODEL: Commerce, not charity
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center bg-accent">
-          <div className="container mx-auto px-4">
-            <div className="max-w-4xl mx-auto text-center">
-              <p className="text-sm uppercase tracking-widest text-accent-foreground/50 mb-4">The Model</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-accent-foreground mb-4 leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                Commerce, not charity
-              </h2>
-              <p className="text-lg text-accent-foreground/70 mb-12 max-w-xl mx-auto leading-relaxed">
-                Products that are Aboriginal owned and controlled, sold commercially.
-                The goal is a community that can come to own the making.
-              </p>
-              <div className="grid gap-4 md:grid-cols-3 mb-12">
-                <div className="bg-accent-foreground/10 rounded-2xl p-8">
-                  <div className="text-5xl font-bold text-accent-foreground mb-2">100%</div>
-                  <div className="text-accent-foreground/70">community ownership is the end goal</div>
-                </div>
-                <div className="bg-accent-foreground/10 rounded-2xl p-8">
-                  <div className="text-5xl font-bold text-accent-foreground mb-2">$0</div>
-                  <div className="text-accent-foreground/70">licensing fees: the model is built to transfer capability to communities over time</div>
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            11. CLIFF PLUMMER: Community testimony video
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center bg-foreground text-background">
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-background/40 mb-4 text-center">
-                Hear From Community
-              </p>
-              <h2
-                className="text-2xl md:text-3xl font-light text-center mb-3"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                {videoTestimonials[0]?.title || 'Community Voices'}
-              </h2>
-              <p className="text-background/50 text-center mb-8">
-                {videoTestimonials[0]?.person} &middot; {videoTestimonials[0]?.location}
-              </p>
-              <VideoSlot
-                src={STORY_VIDEOS.cliffPlummer}
-                label="Community testimony on beds and dignity"
-              />
-              <p className="text-xs text-background/30 mt-4 text-center">
-                {videoTestimonials[0]?.description || 'Video coming soon'}
-              </p>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            11b. FRED CAMPBELL: Oonchiumpa community voice
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-primary mb-4 text-center">
-                Hear From Community
-              </p>
-              <h2
-                className="text-2xl md:text-3xl font-light text-foreground text-center mb-3"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                Fred, Oonchiumpa
-              </h2>
-              <p className="text-muted-foreground text-center mb-8 max-w-lg mx-auto">
-                Fred from the Oonchiumpa Bloomfield family on why community-led manufacturing matters and the path to ownership.
-              </p>
-              <div className="aspect-video rounded-xl overflow-hidden bg-black">
-                <iframe
-                  src="https://share.descript.com/embed/YQwAcYfxzkn"
-                  className="w-full h-full"
-                  allowFullScreen
-                />
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            12. DIANNE STOKES: The washing machine, in her words
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center bg-background">
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-primary mb-4 text-center">
-                The Washing Machine Story
-              </p>
-              <h2
-                className="text-2xl md:text-3xl font-light text-foreground text-center mb-3"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                Elder Dianne Stokes
-              </h2>
-              <p className="text-muted-foreground text-center mb-8 max-w-lg mx-auto">
-                She named it &ldquo;Pakkimjalki Kari&rdquo; in Warumungu language. She designed it
-                in community, tested it, and named it for her family and her Country.
-              </p>
-              {STORY_VIDEOS.dianneStokes ? (
-                <VideoSlot
-                  src={STORY_VIDEOS.dianneStokes}
-                  label="Dianne Stokes on the Pakkimjalki Kari washing machine"
-                />
-              ) : media.people.dianneStokes ? (
-                <div className="relative w-full aspect-video rounded-xl overflow-hidden">
-                  <Image
-                    src={media.people.dianneStokes}
-                    alt="Elder Dianne Stokes"
-                    fill
-                    className="object-cover"
-                    sizes="(max-width: 768px) 100vw, 768px"
-                  />
-                </div>
-              ) : null}
-              <blockquote className="mt-8 text-center">
-                <p
-                  className="text-lg text-foreground leading-relaxed"
-                  style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                >
-                  &ldquo;{codesignQuote?.text}&rdquo;
-                </p>
-                <footer className="flex items-center justify-center gap-3 mt-4">
-                  {avatarMap[codesignQuote?.author ?? ''] ? (
-                    <Image
-                      src={avatarMap[codesignQuote!.author]}
-                      alt={codesignQuote!.author}
-                      width={40}
-                      height={40}
-                      className="rounded-full object-cover w-10 h-10"
-                    />
-                  ) : (
-                    <div className="w-10 h-10 rounded-full bg-muted flex items-center justify-center text-muted-foreground font-medium">
-                      {codesignQuote?.author?.charAt(0)}
-                    </div>
-                  )}
-                  <div className="text-left text-sm">
-                    <span className="font-medium text-foreground">{codesignQuote?.author}</span>
-                    <br />
-                    <span className="text-muted-foreground">{codesignQuote?.context}</span>
-                  </div>
-                </footer>
-              </blockquote>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            13. PARTNERS: Who makes this possible
-            ════════════════════════════════════════════════════════════ */}
-        <section id="partners" className="min-h-screen flex items-center justify-center" style={{ backgroundColor: '#FDF8F3' }}>
-          <div className="container mx-auto px-4">
-            <div className="max-w-5xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-primary mb-4 text-center">Partners</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-foreground mb-12 text-center leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                This work doesn&rsquo;t happen alone
-              </h2>
-              <div className="grid gap-8 md:grid-cols-2 lg:grid-cols-3 mb-12">
-                <div className="rounded-2xl border border-border bg-background p-6">
-                  <h3 className="font-semibold text-foreground mb-3 text-sm uppercase tracking-wide">Community Partners</h3>
-                  <ul className="space-y-2 text-sm text-muted-foreground">
-                    {partners.communityPartners.map((p) => (
-                      <li key={p.name}>
-                        <span className="font-medium text-foreground">{p.name}</span>
-                        <br />
-                        <span className="text-xs">{p.role} &middot; {p.location}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-                <div className="rounded-2xl border border-border bg-background p-6">
-                  <h3 className="font-semibold text-foreground mb-3 text-sm uppercase tracking-wide">Health Partners</h3>
-                  <ul className="space-y-2 text-sm text-muted-foreground">
-                    {partners.healthPartners.map((p) => (
-                      <li key={p.name}>
-                        <span className="font-medium text-foreground">{p.name}</span>
-                        <br />
-                        <span className="text-xs">{p.location}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-                <div className="rounded-2xl border border-border bg-background p-6">
-                  <h3 className="font-semibold text-foreground mb-3 text-sm uppercase tracking-wide">Funders &amp; Supporters</h3>
-                  <ul className="space-y-2 text-sm text-muted-foreground">
-                    {partners.fundingPartners.map((p) => (
-                      <li key={p.name}>
-                        <span className="font-medium text-foreground">{p.name}</span>
-                        <br />
-                        <span className="text-xs">{p.type}</span>
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              </div>
-              {/* Advisory board */}
-              <div className="text-center">
-                <h3 className="text-sm uppercase tracking-widest text-muted-foreground mb-6">Advisory Group</h3>
-                <div className="flex flex-wrap justify-center gap-3">
-                  {advisoryGroup.map((person) => (
-                    <div key={person.name} className="px-4 py-2 rounded-full border border-border bg-background text-sm">
-                      <span className="font-medium text-foreground">{person.name}</span>
-                      {person.org && <span className="text-muted-foreground"> &middot; {person.org}</span>}
-                    </div>
-                  ))}
-                </div>
-              </div>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            14. WHAT'S NEXT: Vision and scale
-            ════════════════════════════════════════════════════════════ */}
-        <section id="future" className="min-h-screen flex items-center justify-center relative overflow-hidden">
-          <video
-            className="absolute inset-0 w-full h-full object-cover"
-            autoPlay
-            muted
-            loop
-            playsInline
-            poster={STORY_VIDEOS.buildingTogether.poster}
-          >
-            <source src={STORY_VIDEOS.buildingTogether.desktop} media="(min-width: 768px)" type="video/mp4" />
-            <source src={STORY_VIDEOS.buildingTogether.mobile} type="video/mp4" />
-          </video>
-          <div className="absolute inset-0 bg-black/70" />
-          <div className="relative z-10 container mx-auto px-4">
-            <div className="max-w-4xl mx-auto">
-              <p className="text-sm uppercase tracking-widest text-white/40 mb-4 text-center">What&rsquo;s Next</p>
-              <h2
-                className="text-3xl md:text-5xl font-light text-white mb-12 text-center leading-snug"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                Containerised manufacturing at scale
-              </h2>
-              <div className="grid gap-4 md:grid-cols-3 mb-12">
-                {[
-                  { year: 'Year 1 (2026–27)', place: 'QLD Flagship: Jinibara Country', stat: '1,500 beds · 6 jobs' },
-                  { year: 'Year 2 (2027–28)', place: 'Central Australia: Oonchiumpa', stat: '3,500 beds · 12 jobs' },
-                  { year: 'Year 3 (2028–29)', place: 'Top End or Torres Strait', stat: '5,000 beds · 18 jobs · 125t plastic diverted' },
-                ].map((phase) => (
-                  <div key={phase.year} className="rounded-xl bg-white/5 border border-white/10 p-6 text-center">
-                    <div className="text-xs uppercase tracking-widest text-primary mb-2">{phase.year}</div>
-                    <h3 className="font-semibold text-white mb-2">{phase.place}</h3>
-                    <p className="text-sm text-white/50">{phase.stat}</p>
-                  </div>
-                ))}
-              </div>
-              <blockquote className="text-center">
-                <p
-                  className="text-2xl md:text-3xl font-light text-white leading-relaxed"
-                  style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-                >
-                  &ldquo;When someone asks &lsquo;Who makes these?&rsquo; and the answer is &lsquo;We do.&rsquo;&rdquo;
-                </p>
-              </blockquote>
-            </div>
-          </div>
-        </section>
-
-        {/* ════════════════════════════════════════════════════════════════
-            15. THE INVITATION: CTA
-            ════════════════════════════════════════════════════════════ */}
-        <section className="min-h-screen flex items-center justify-center bg-foreground text-background">
-          <div className="container mx-auto px-4">
-            <div className="max-w-3xl mx-auto text-center">
-              <h2
-                className="text-3xl md:text-5xl font-light leading-snug mb-8"
-                style={{ fontFamily: 'var(--font-display, Georgia, serif)' }}
-              >
-                We&rsquo;re not building beds.{' '}
-                <span className="text-primary">We&rsquo;re building a model where communities manufacture their own future.</span>
-              </h2>
-              <p className="text-lg text-background/50 mb-12 max-w-xl mx-auto leading-relaxed">
-                Durable products from community waste. Local jobs. Community ownership.
-                The disposable furniture cycle ends here.
-              </p>
-              <div className="grid gap-4 sm:grid-cols-3 max-w-2xl mx-auto mb-8">
-                <Link href="/shop/stretch-bed-single" className="group rounded-2xl border border-white/20 p-6 text-center hover:bg-white/5 transition-colors">
-                  <div className="text-3xl mb-2">🛏️</div>
-                  <h3 className="font-semibold text-white mb-1">Buy a Stretch Bed</h3>
-                  <p className="text-xs text-white/40">From $750</p>
-                </Link>
-                <Link href="/sponsor" className="group rounded-2xl border border-primary/40 bg-primary/5 p-6 text-center hover:bg-primary/10 transition-colors">
-                  <div className="text-3xl mb-2">💛</div>
-                  <h3 className="font-semibold text-primary mb-1">Sponsor a Bed</h3>
-                  <p className="text-xs text-white/40">Gift a bed to community</p>
-                </Link>
-                <Link href="/partner" className="group rounded-2xl border border-white/20 p-6 text-center hover:bg-white/5 transition-colors">
-                  <div className="text-3xl mb-2">🤝</div>
-                  <h3 className="font-semibold text-white mb-1">Partner With Us</h3>
-                  <p className="text-xs text-white/40">Organisations &amp; funders</p>
-                </Link>
-              </div>
-              <p className="text-sm text-white/30">
-                hi@act.place &middot; goodsoncountry.com
-              </p>
-            </div>
-          </div>
-        </section>
-
       </div>
 
-      {/* Utility + print styles */}
-      <style>{`
-        .no-scrollbar::-webkit-scrollbar { display: none; }
-        .no-scrollbar { -ms-overflow-style: none; scrollbar-width: none; }
-        @media print {
-          nav { display: none !important; }
-          .story-scroll section {
-            min-height: auto !important;
-            background: white !important;
-            color: black !important;
-            padding: 1.5rem 0 !important;
-            page-break-inside: avoid;
-          }
-          .story-scroll h1, .story-scroll h2, .story-scroll h3,
-          .story-scroll p, .story-scroll li, .story-scroll span,
-          .story-scroll blockquote, .story-scroll footer {
-            color: black !important;
-          }
-          .story-scroll video, .story-scroll iframe {
-            display: none !important;
-          }
-        }
-      `}</style>
+      {stop.voiceNames?.map((name) => (
+        <Voice key={name} name={name} />
+      ))}
+
+      {stop.figures && stop.figures.length > 0 && <Figures figures={stop.figures} />}
+      {stop.gallery && stop.gallery.length > 0 && <Gallery images={stop.gallery} />}
+
+      {stop.videos?.map((video) => (
+        <Video key={video.kind === 'local' ? video.src : video.embedUrl} video={video} />
+      ))}
+
+      {showGaps &&
+        stop.gaps?.map((gap) => <Gap key={`${gap.slot}-${gap.wanted}`} gap={gap} />)}
+    </section>
+  );
+}
+
+// ─────────────────────────────────────────────────────────── page
+
+export default async function StoryRoadPage({
+  searchParams,
+}: {
+  searchParams: Promise<{ review?: string }>;
+}) {
+  const placeStops = storyStops.filter((s) => s.kind === 'stop');
+  const { review } = await searchParams;
+  // Review mode: everywhere in development, on ?review=1 in production.
+  const showGaps = process.env.NODE_ENV === 'development' || review === '1';
+
+  return (
+    <main className="mx-auto max-w-3xl px-5 pb-32 pt-16 md:pt-24">
+      <header>
+        <p className="text-sm uppercase tracking-[0.14em] text-muted-foreground">
+          The Work That Stays
+        </p>
+        <h1 className="mt-4 font-serif text-4xl md:text-6xl leading-[1.05] text-foreground">
+          {storyOpening.headline}
+        </h1>
+
+        <figure className="my-12">
+          <Image
+            src={storyOpening.photo.src}
+            alt={storyOpening.photo.alt}
+            width={1400}
+            height={900}
+            sizes="(max-width: 768px) 100vw, 70vw"
+            priority
+            className="w-full h-auto rounded-md object-cover"
+          />
+          <figcaption className="mt-2 text-sm text-muted-foreground">
+            {storyOpening.photo.caption}
+          </figcaption>
+        </figure>
+
+        <div className="space-y-6">
+          {storyOpening.chapters.map((para, i) => (
+            <p key={i} className="text-lg md:text-xl leading-relaxed text-foreground/90">
+              {para}
+            </p>
+          ))}
+        </div>
+      </header>
+
+      {/* The road, listed before it is walked. Seven places, then what they produced. */}
+      <nav aria-label="The road" className="mt-16 border-t pt-8">
+        <p className="text-sm uppercase tracking-[0.14em] text-muted-foreground">The road</p>
+        <ol className="mt-4 space-y-2">
+          {placeStops.map((stop, i) => (
+            <li key={stop.id} className="text-base">
+              <Link href={`#${stop.id}`} className="hover:underline text-foreground">
+                <span className="text-muted-foreground tabular-nums mr-3">{i + 1}</span>
+                {stop.headline}
+              </Link>
+              {stop.place && (
+                <span className="block ml-8 text-sm text-muted-foreground">{stop.place}</span>
+              )}
+            </li>
+          ))}
+        </ol>
+      </nav>
+
+      {storyStops.map((stop) => (
+        <Stop key={stop.id} stop={stop} showGaps={showGaps} />
+      ))}
+
+      <footer className="border-t pt-10 text-sm text-muted-foreground">
+        <p>Updated {storyUpdated}.</p>
+        <p className="mt-2">
+          Figures are read from the canonical register. Each carries its own claims
+          status, because this page mixes delivered counts with modelled economics and
+          the two must not read alike.
+        </p>
+      </footer>
     </main>
   );
 }
