+++ wiki/canon/needs-signoff.md
@@ -1,4 +1,4 @@
-# Canon money external-send checks (as of 2026-08-01)
+# Canon money external-send checks (as of 2026-08-02)
 
 Generated by scripts/check-canon-drift.mjs. Money is never auto-written; re-pull and clear these by hand before an external send.
 
@@ -15,7 +15,7 @@ Still check before sending: opening cash, AP matching clean-up, latest receipts,
 None. Money figures are in lockstep across canon.ts, compendium.ts, grant-content.ts.
 
 ## Staleness warnings
-- revenueReceived (revenueReceived) is 59 days old (asAt 2026-06-03). Re-pull Xero and reconcile before any external share.
-- accountsReceivable (accountsReceivable) is 59 days old (asAt 2026-06-03). Re-pull Xero and reconcile before any external share.
-- xeroPaid (ACT-GD receivables paid (Xero cut)) is 61 days old (asAt 2026-06-01). Re-pull Xero and reconcile before any external share.
-- carveOut (Goods revenue carve-out) is 60 days old (asAt 2026-06-02). Re-pull Xero and reconcile before any external share.
+- revenueReceived (revenueReceived) is 60 days old (asAt 2026-06-03). Re-pull Xero and reconcile before any external share.
+- accountsReceivable (accountsReceivable) is 60 days old (asAt 2026-06-03). Re-pull Xero and reconcile before any external share.
+- xeroPaid (ACT-GD receivables paid (Xero cut)) is 62 days old (asAt 2026-06-01). Re-pull Xero and reconcile before any external share.
+- carveOut (Goods revenue carve-out) is 61 days old (asAt 2026-06-02). Re-pull Xero and reconcile before any external share.
