+++ wiki/canon/story-coverage.md
@@ -1,6 +1,6 @@
 # Story + illustration coverage (Loop E)
 
-Generated by v2/scripts/check-story-coverage.mjs on 2026-08-01. Read-only. Loop E does not generate anything: it tells the /ledger-story and /goods-illustrations skills what to draft next. Generation is human-triggered, drafts only, consent re-gated each run.
+Generated by v2/scripts/check-story-coverage.mjs on 2026-08-02. Read-only. Loop E does not generate anything: it tells the /ledger-story and /goods-illustrations skills what to draft next. Generation is human-triggered, drafts only, consent re-gated each run.
 
 Sources: the cleared-voice pool = curated-quotes.ts keys ∪ trip-stories.ts VoiceCards with `consent: 'cleared'` (mirrors ledger-story/CONSENT.md sources 1-2; EL live check is out of scope for a hermetic script). Weekly coverage = front-matter `storyteller:` in wiki/outputs/ledger/*.md. Illustration coverage = goods-ill-*.png in v2/public/images/brand vs draft sets in generated-images/goods-illustrations/. RED-safe: counts + already-public display_names only, never quotes or recipient data.
 
@@ -9,7 +9,7 @@ Every weekly ledger draft features a storyteller on the cleared-voices allowlist
 
 ## Weekly cadence
 
-🟢 On track: -23 days since the last post (2026-08-24).
+🟢 On track: -22 days since the last post (2026-08-24).
 
 ## Cleared-voice pool
